% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_vis.R
\name{plot_dim}
\alias{plot_dim}
\title{Title}
\usage{
plot_dim(
  object,
  group.by = NULL,
  reduction = NULL,
  show.pt = TRUE,
  pt.size = 0.5,
  fill = NA,
  col = NULL,
  show.ct = TRUE,
  show.cls = c("All"),
  label.size = 5
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{group.by}{Group (color) cells in different ways (for example, celltype)}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap}

\item{show.pt}{Either to show points, default TRUE, FALSE means do not show points}

\item{pt.size}{Point size for points}

\item{fill}{NA means not fill the edge line contour, TRUE means fill}

\item{col}{Colors to use for plotting}

\item{show.ct}{Either to show edge line contour, default TRUE (show), if FALSE (now show)}

\item{show.cls}{A vector, default c('All'), can be subset of group.by clusters}

\item{label.size}{label size}
}
\value{
A \code{\link[patchwork]{patchwork}ed} ggplot object
}
\description{
Title
}
\examples{
plot_dim(rds, fill=TRUE)
plot_dim(rds, fill=NA, show.cls=c("CD4 Naive T", "CD14 Mono"))
plot_dim(rds, fill=TRUE, show.cls=c("T activated", "CD14 Mono"))
plot_dim(rds, fill=NA)
plot_dim(rds, fill=TRUE, show.pt=FALSE)
plot_dim(rds, fill=NA, show.pt=FALSE)
plot_dim(rds, reduction="pca")

}
